Liten refactor av prune-backend-node-modules og endrer extension fra .mjs til .js
